+++ carp_mobile_sensing/lib/probes/connectivity/connectivity.dart
@@ -10,7 +10,6 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:connectivity/connectivity.dart';
 import 'package:carp_mobile_sensing/core/core.dart';
 
-part 'connectivity_probe.dart';
-part 'bluetooth_probe.dart';
+part 'connectivity_probes.dart';
 part 'connectivity_datum.dart';
 part 'connectivity.g.dart';