+++ carp_mobile_sensing/lib/probes/device/device.dart
@@ -1,8 +1,9 @@
 /// A library for collecting information from the device hardware:
-/// - battery
-/// - screen
-/// - free memory
-library hardware;
+///  - device info
+///  - battery
+///  - screen
+///  - free memory
+library device;
 
 import 'dart:async';
 import 'package:battery/battery.dart';
@@ -12,7 +13,6 @@ import 'package:screen_state/screen_state.dart';
 import 'package:system_info/system_info.dart';
 import 'package:carp_mobile_sensing/core/core.dart';
 
-part 'battery_probe.dart';
-part 'hardware_datum.dart';
-part 'hardware_probes.dart';
-part 'hardware.g.dart';
+part 'device_datum.dart';
+part 'device_probes.dart';
+part 'device.g.dart';